update test scripts

+++ tests/testvm.sh
@@ -11,7 +11,7 @@ flag=${2:-R}
 gethostdata ${host}
 
 if [[ ${type} != vm && $type != vmlocal ]]; then
-  echo "-- $(date '+%T') ${host}: not a vm"
+  echo "-- $(TZ=PST8PDT date '+%T') ${host}: not a vm"
   exit 1
 fi
 
@@ -19,11 +19,11 @@ already=F
 ./tests/startvm.sh ${host} T
 rc=$?
 if [[ $rc -ne 0 && $rc -ne 255 ]]; then
-  echo "-- $(date '+%T') ${host}: unable to start"
+  echo "-- $(TZ=PST8PDT date '+%T') ${host}: unable to start"
   exit 1
 fi
 if [[ $rc -eq 255 ]]; then
-  echo "-- $(date '+%T') ${host}: already started"
+  echo "-- $(TZ=PST8PDT date '+%T') ${host}: already started"
   already=T
 fi
 
@@ -36,19 +36,21 @@ if [[ x${ipaddr} == "-" ]]; then
   exit 1
 fi
 
-tarfn=$(echo di-*.tar.gz)
-didir=$(echo ${tarfn} | sed 's,\.tar.gz$,,')
+if [[ $flag != "CHK" ]]; then
+  tarfn=$(echo di-*.tar.gz)
+  didir=$(echo ${tarfn} | sed 's,\.tar.gz$,,')
 
-rsltdir=$(pwd)/test_results/${host}
-test -d ${rsltdir} && rm -rf ${rsltdir}
-mkdir -p ${rsltdir}
+  rsltdir=$(pwd)/test_results/${host}
+  test -d ${rsltdir} && rm -rf ${rsltdir}
+  mkdir -p ${rsltdir}
 
-./tests/thost.sh ${tarfn} ${didir} ${host} ${type} \
-    ${ipaddr} ${remuser} ${remport} ${rempath} \
-    ${flag} ${complist} 2>&1 | tee ${rsltdir}/w
+  ./tests/thost.sh ${tarfn} ${didir} ${host} ${type} \
+      ${ipaddr} ${remuser} ${remport} ${rempath} \
+      ${flag} ${complist} 2>&1 | tee ${rsltdir}/w
+fi
 
 if [[ $already == F ]]; then
-  echo "-- $(date '+%T') ${host}: stopping"
+  echo "-- $(TZ=PST8PDT date '+%T') ${host}: stopping"
   ./tests/stopvm.sh ${host}
 fi
